refactor(): have landing be a simple render function
Moves landing into helper function
