+++ src/helpers/index.js
@@ -7,6 +7,7 @@ import {material} from 'util'
 import 'images/sn-logo-32.png'
 
 export {menu} from './menu'
+export {landing} from './landing'
 
 // helper function to attach react components to the snabbdom
 // some need to be attached on 'update', others on 'insert', not sure why
